requirements.txt
streamlit==1.24.0
numpy==1.24.3
pillow
tensorflow-cpu==2.13.0
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
pandas==2.2.0
matplotlib==3.8.0
googletrans==4.0.0-rc1
pyttsx3==2.90
boto3==1.28.0
python-dotenv==1.0.0
twilio==8.0.0
scikit-learn==1.3.0
gunicorn==20.1.0
redis==4.5.5
python-multipart
black
